Rename 'scenario helpers' to 'repository', and move prisma setup file to server dir
